switch to integers

+++ Limit.cpp
@@ -1,6 +1,6 @@
 #include "Limit.h"
 
-double Limit::getPrice() {
+int Limit::getPrice() {
     return price;
 }
 
@@ -8,7 +8,7 @@ std::queue<Order>& Limit::getOrders() {
     return orders;
 }
 
-Limit::Limit(double price) : price(price) {};
+Limit::Limit(int price) : price(price) {};
 
 void Limit::addOrder(Order& order) {
     orders.push(order);
